+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `Detections.draw_on` no longer breaks if some classes are unknown (i.e. null)
 * Fix overflow issue in `py_nms` if inputs were float16
 * Fixed issue preventing torchvision nms from being recognized as a backend.
+* Issue with Points.coerce when data was a numpy array
 
 ### Changed
 * kwimage.Affine.fit now falls back to a similarity transform if fewer than 3 correspondences are given,